+++ lib/local_country_picker.dart
@@ -3,6 +3,7 @@
 library;
 
 export 'src/country.dart';
+export 'src/country_flag.dart';
 export 'src/country_flag_shape.dart';
 export 'src/country_picker_localizations.dart';
 export 'src/country_picker_theme.dart';